Move setOmedaData and refactor/doc
