+++ packages/marko-web-omeda-identity-x/integration-hooks/on-login-link-sent.js
@@ -1,44 +1,13 @@
-const gql = require('graphql-tag');
-const { get } = require('@parameter1/base-cms-object-path');
 const extractPromoCode = require('../utils/extract-promo-code');
 const {
   getOmedaCustomerRecord,
   getOmedaLinkedFields,
+  setOmedaData,
   setOmedaDemographics,
   setOmedaDeploymentTypes,
+  getAnsweredQuestionMap,
 } = require('../omeda-data');
 
-const SET_OMEDA_DATA = gql`
-  mutation SetOmedaData($input: SetAppUserUnverifiedDataMutationInput!) {
-    setAppUserUnverifiedData(input: $input) { id }
-  }
-`;
-
-/**
- * Sets Omeda customer data (such as name, address, etc) to the IdentityX user fields.
- */
-const setOmedaData = async ({ identityX, user, omedaCustomer }) => {
-  const input = {
-    email: user.email,
-
-    givenName: omedaCustomer.firstName,
-    familyName: omedaCustomer.lastName,
-    organization: omedaCustomer.companyName,
-    organizationTitle: omedaCustomer.title,
-
-    countryCode: get(omedaCustomer, 'primaryPostalAddress.countryCode'),
-    regionCode: get(omedaCustomer, 'primaryPostalAddress.regionCode'),
-    postalCode: get(omedaCustomer, 'primaryPostalAddress.postalCode'),
-    street: get(omedaCustomer, 'primaryPostalAddress.street'),
-    addressExtra: get(omedaCustomer, 'primaryPostalAddress.extraAddress'),
-    city: get(omedaCustomer, 'primaryPostalAddress.city'),
-    phoneNumber: get(omedaCustomer, 'primaryPhoneNumber.phoneNumber'),
-  };
-  return identityX.client.mutate({
-    mutation: SET_OMEDA_DATA,
-    variables: { input },
-  });
-};
 
 module.exports = async ({
   brandKey,
@@ -73,25 +42,14 @@ module.exports = async ({
     }),
   ]);
 
-  const answeredQuestionMap = new Map();
-  user.customSelectFieldAnswers.forEach((select) => {
-    if (!select.hasAnswered) return;
-    answeredQuestionMap.set(select.field.id, true);
-  });
-  user.customBooleanFieldAnswers.forEach((boolean) => {
-    if (!boolean.hasAnswered) return;
-    answeredQuestionMap.set(boolean.field.id, true);
-  });
-
+  const answeredQuestionMap = getAnsweredQuestionMap(user);
   const hasAnsweredAllOmedaQuestions = omedaLinkedFields
     .demographic.every(field => answeredQuestionMap.has(field.id))
     && omedaLinkedFields
       .deploymentType.every(field => answeredQuestionMap.has(field.id));
 
-  // @todo: Check if a valid answer exists for all omeda questions
-  // @todo: user.hasAnsweredAllOmedaQuestions isn't a thing, review this.
-  // Possible typo for hasAnsweredAllOmedaQuestions to bypass omeda fetch/updates?
-  if (user.verified && user.hasAnsweredAllOmedaQuestions) {
+  // If the user is verified and has answered all Omeda questions, do nothing.
+  if (user.verified && hasAnsweredAllOmedaQuestions) {
     return;
   }
 
@@ -100,6 +58,8 @@ module.exports = async ({
 
   // Only set existing omeda data to IdentityX if the user is not yet verified (aka, new user)
   if (!user.verified) await setOmedaData({ identityX, user, omedaCustomer });
+
+  // If the IdX user is missing answers, set from existing omeda demos (custom select/bool fields)
   if (!hasAnsweredAllOmedaQuestions) {
     await setOmedaDemographics({
       identityX,
