+++ packages/marko-web-omeda-identity-x/omeda-data/index.js
@@ -1,6 +1,7 @@
 const getAnsweredQuestionMap = require('./get-answered-question-map');
 const getOmedaCustomerRecord = require('./get-omeda-customer-record');
 const getOmedaLinkedFields = require('./get-omeda-linked-fields');
+const setOmedaData = require('./set-omeda-data');
 const setOmedaDemographics = require('./set-omeda-demographics');
 const setOmedaDeploymentTypes = require('./set-omeda-deployment-types');
 
@@ -8,6 +9,7 @@ module.exports = {
   getAnsweredQuestionMap,
   getOmedaCustomerRecord,
   getOmedaLinkedFields,
+  setOmedaData,
   setOmedaDemographics,
   setOmedaDeploymentTypes,
 };
